solution-2/creature-generator/requirements.txt
flask[async]==2.3.2
Pillow
requests
gunicorn==20.1.0
python-dotenv

